Add audio item to ConversationList and remove number input from NewConversation

+++ src/components/ConversationList/item.tsx
@@ -1,6 +1,6 @@
 import { Conversation } from "@/models";
 import { TextBody } from "../MessageBalloon/text-body";
-import { Camera, File } from "lucide-react";
+import { Camera, File, Mic } from "lucide-react";
 
 type MessageItemProps = {
   conversation: Conversation;
@@ -11,6 +11,8 @@ export const MessageItem = ({ conversation, className }: MessageItemProps) => {
   if (conversation?.type == "image") return <ImageItem />;
 
   if (conversation?.type == "document") return <DocumentItem />;
+
+  if (conversation?.type == "audio") return <AudioItem />;
 
   return (
     <span className={`text-[#aebac1] text-sm truncate ${className}`}>
@@ -33,6 +35,15 @@ const DocumentItem = () => {
     <span className="text-[#aebac1] text-sm truncate flex items-center gap-1">
       <File className="w-[14px]" />
       Documento
+    </span>
+  );
+};
+
+const AudioItem = () => {
+  return (
+    <span className="text-[#aebac1] text-sm truncate flex items-center gap-1">
+      <Mic className="w-[14px]" />
+      Audio
     </span>
   );
 };
